chore(): refactor names on navigation
